BE/Fix : Claude Code #1 Review 반영
- long타입으로 구현된 부분들 중 꼭 필요한 부분을 제외하고 double로 변경

+++ backend/src/main/java/com/naengsam/quick/domain/matching/model/MatchingCandidate.java
@@ -10,7 +10,7 @@ import java.util.UUID;
 public record MatchingCandidate(
         UUID orderId,
         UUID dreamiId,
-        long distanceMeters,
+        double distanceMeters,
         Duration orderWaitingTime,
         Duration dreamiWaitingTime,
         int orderCandidateCount,
